requirements.txt
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bcrypt==4.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
fonttools==4.54.1
html5lib==1.1
idna==3.9
IronPdf==2024.9.1.3
lxml==5.3.0
oscrypto==1.3.0
paypalrestsdk==1.13.3
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pyOpenSSL==24.2.1
pypdf==5.0.0
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dotenv==1.0.1
pythonnet==3.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.11.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
zopfli==0.2.3
